requirements.txt
flask>=1.1.1
flask-sqlalchemy>=2.4.0
psycopg2-binary>=2.8.3
alembic>=1.2.0
flask-wtf>=0.14.2
WTForms>=2.2.1
flask-bootstrap>=3.3.7.1
SQLAlchemy>=1.3.8
Pillow>=6.1.0
bcrypt>=3.1.7